if ehpa does not contain predictable metrics, then don't create a tsp object
